init: skeleton book layout and tooling
I am using latex to write the book as it renders beautiful layouts. I'd ideally
like to produce a web version also but I am deferring thinking about exactly how
I'll do this.

I have added a first guess at chapters, a first attempt at a preface and the
beginnings of an introduction.

For easier dissemination I will be including the generated PDF in the repo.